(FreeCAD module) Brep.hs: fixed some bug that would make divisions by 0 if the read object was not rotated. git-svn-id: https://svn-agbkb.informatik.uni-bremen.de/Hets/trunk@15112 cec4b9c1-7d33-0410-9eda-942365e851bb

+++ FreeCAD/brep_conversion/src/brep_to_xml.cxx
@@ -297,10 +297,10 @@ void BrepToXML::build_xml(const string& what, string& output)
     stringstream ss;
     ss    << "<rectangle>\n";
     vector <int> duplicates;
-    for (int i = 0; i < vLocs.size(); i++)
+    for (int i = 0; i < (int) vLocs.size(); i++)
       {
 	bool write = true;
- 	for (int j = 0; j < duplicates.size(); j++)
+ 	for (int j = 0; j < (int) duplicates.size(); j++)
    	  {
 	    if ( vLocs[i].first == duplicates[j])
 	      write = false;
